- Snake::get_ai_inputs reworked: floats and correct distances - Added new c-tor for Snake: initial body length is 3 items - TODO: fix bug in Population c-tors

+++ snake-ai-sfml/ai_agent/include/Matrix.h
@@ -31,7 +31,7 @@ public:
 	/*!
 	* \brief Creates a single column matrix
 	*/
-	Matrix( vector<int>& inputs );
+	Matrix( vector<float>& inputs );
 
 	/*!
 	* \brief creates from Eigen MatrixXd
